Prevent the control clear patch from blocking ctrl-u
The control clear patch is not compatible with the latest ctrl-u feature
because it blocks all ctrl shortcuts. This fix ensures that the patch is
active only when the ctrl key is pressed without any other keys.

Note that the patch is also active when the shift or alt key is pressed
without any other keys. This is an undocumented feature of this patch
and is caused by the iscntrl function, which returns true if the ascii
value is zero.